+++ det.h
@@ -1,7 +1,7 @@
 #ifndef DET_H
 #define DET_H
 
-long long det(int n, int **mat);
+int det(int n, int **mat);
 int gcd(int a, int b);
 
 #endif
